requirements.txt
Django >=4.0, <5.0
mysqlclient
python-dotenv
djangorestframework-simplejwt
djangorestframework
django-rest-swagger
pyyaml
Pillow

redgreenunittest
coverage
